pyproject.toml
# Travel Planner MCP Server
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-planner-mcp"
version = "0.2.1"
description = "여행 숙소, 항공편, 관광지를 검색하고 추천하는 MCP 서버 (Booking.com API)"
requires-python = ">=3.10"
license = "MIT"
readme = "README.md"
authors = [{ name = "idean3885" }]
keywords = ["travel", "booking", "mcp", "claude-code", "hotels", "flights", "attractions"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
travel-planner-mcp = "travel_mcp.server:main"

[project.urls]
Homepage = "https://github.com/idean3885/travel-planner"
Repository = "https://github.com/idean3885/travel-planner"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "mcp-servers/hotels_mcp_server"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.setuptools.packages.find]
where = ["src"]
